+++ packages/identicon/index.ts
@@ -2,4 +2,4 @@ export { default as IdenticonBadge } from './IdenticonBadge'
 export * from './IdenticonBadge'
 export { default as Identicon } from './Identicon'
 export * from './Identicon'
-export * from './types'
+export * from './constants'
