+++ packages/input/types.ts
@@ -1,35 +1,5 @@
 import { ReactNode, RefObject } from 'react'
-export type { Theme } from '@lidofinance/theme'
-
-export enum InputMessageVariant {
-  error,
-  warning,
-  success,
-}
-export type InputMessageVariants = keyof typeof InputMessageVariant
-
-export enum InputType {
-  text,
-  password,
-  number,
-  email,
-  search,
-  tel,
-  url,
-}
-export type InputTypes = keyof typeof InputType
-
-export enum InputVariant {
-  small,
-  default,
-}
-export type InputVariants = keyof typeof InputVariant
-
-export enum InputColor {
-  default,
-  accent,
-}
-export type InputColors = keyof typeof InputColor
+import { InputVariants, InputColors } from './constants'
 
 export type CommonProps = {
   label?: ReactNode
