+++ packages/select/types.ts
@@ -1,4 +1,2 @@
-export type { Theme } from '@lidofinance/theme'
-
 export type OptionValue = string | number
 export type SelectHandleChange = (value: OptionValue) => void
